Move test-helpers out
