+++ client/layouts/specs/LogoutLayout.spec.tsx
@@ -4,13 +4,13 @@ import { HashRouter, Route, Redirect } from 'react-router-dom';
 
 // ----- Test Helpers
 import '@testing-library/jest-dom/extend-expect';
-import renderWithRouter from '../test-helpers/router.helper.spec';
+import renderWithRouter from '../../test-helpers/router.helper.spec';
 import { render, cleanup, fireEvent, waitForElement } from '@testing-library/react';
 
 import configureStore from '../../store/configureStore'
 
 import LogoutLayout from '../Logout/LogoutLayout'
-import ReduxWrap from '../test-helpers/ReduxWrap.helper.spec';
+import ReduxWrap from '../../test-helpers/ReduxWrap.helper.spec';
 
 import { StoreShape } from '../../types';
