+++ client/layouts/specs/MainLayout/SubmitBar.spec.tsx
@@ -6,7 +6,7 @@ import axios from '../../../axios';
 import '@testing-library/jest-dom/extend-expect';
 import { render, cleanup, fireEvent, waitForElement, act } from '@testing-library/react';
 
-import ReduxWrap from '../../test-helpers/ReduxWrap.helper.spec';
+import ReduxWrap from '../../../test-helpers/ReduxWrap.helper.spec';
 import configureStore from '../../../store/configureStore';
 
 import { StoreShape } from '../../../types';
